backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
supabase==1.0.4
python-dotenv==1.0.0
bcrypt==4.0.1
PyJWT==2.8.0
email-validator==2.1.0
